requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.33.0
langgraph>=0.0.30
langchain-core>=0.1.30
langchain-google-vertexai>=0.1.0
sqlalchemy[asyncio]>=2.0.29
pgvector>=0.2.5
cloud-sql-python-connector[asyncpg]>=1.4.0
pydantic>=2.7.0
python-dotenv>=1.0.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
jira>=3.8.0
pytz>=2024.1
